requirements.txt
selene
pytest
pydantic-settings
pydantic
python-dotenv